Add a way to name native routines
Sometimes the names used in a particular native library may not mesh
well with their usage in Perl. Rather than force our Perl subroutines to
have the same name as their native counterparts, we provide a way to
specify the name of the routine in the native library that our arbitrarily-
named Perl subroutine maps to.