Make the test suite run on JVM.
Here be hackery. Most notably, figuring out how to compile libraries on JVM is
tricky, so for now we can run the test suite with Rakudo/Parrot to get them
built and then again with Rakudo/JVM.

Also, the extension guessing doesn't, so far. It just assumes the shared
library extension is so.